refactor schedules and jobs method
