+++ python_modules/libraries/dagster-airflow/dagster_airflow_tests/test_dagster_job_factory/test_schedule_timezone.py
@@ -2,7 +2,6 @@ import pendulum
 import pytest
 from airflow import __version__ as airflow_version
 from airflow.models.dag import DAG
-from dagster import job
 from dagster_airflow.dagster_schedule_factory import make_dagster_schedule_from_airflow_dag
 
 
@@ -18,10 +17,6 @@ def test_schedule_timezone():
         schedule="0 0 * * *",
     )
 
-    @job
-    def job_def():
-        return
-
-    schedule = make_dagster_schedule_from_airflow_dag(dag=dag, job_def=job_def)
+    schedule = make_dagster_schedule_from_airflow_dag(dag=dag)
     assert schedule.cron_schedule == "0 0 * * *"
     assert schedule.execution_timezone == "Europe/London"
