+++ python_modules/libraries/dagster-airflow/dagster_airflow_tests/test_dagster_pipeline_factory/test_load_connections.py
@@ -7,7 +7,7 @@ from unittest import mock
 import pytest
 from airflow import __version__ as airflow_version
 from airflow.models import Connection
-from dagster_airflow.dagster_job_factory import make_dagster_definitions_from_airflow_dags_path
+from dagster_airflow import make_dagster_definitions_from_airflow_dags_path
 
 from dagster_airflow_tests.marks import requires_airflow_db
