+++ python_modules/libraries/dagster-airflow/dagster_airflow_tests/test_dagster_pipeline_factory/test_solid_execution.py
@@ -25,7 +25,7 @@ from dagster._core.instance import AIRFLOW_EXECUTION_DATE_STR
 from dagster._core.storage.compute_log_manager import ComputeIOType
 from dagster._core.test_utils import instance_for_test
 from dagster._seven import get_current_datetime_in_utc
-from dagster_airflow.dagster_job_factory import make_dagster_job_from_airflow_dag
+from dagster_airflow import make_dagster_job_from_airflow_dag
 
 default_args = {
     "owner": "dagster",
